requirements.txt
kafka-python
pandas
psycopg2-binary
streamlit
plotly
sqlalchemy